# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gold Standard Cleaners internal management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface PasswordUpdate {
  currentPassword: string;
  newPassword: string;
}

export interface AuthUser {
  id: number;
  username: string;
}

export interface DashboardStats {
  revenue: number;
  expenses: number;
  netProfit: number;
  jobCount: number;
  revenueChange: number;
  expensesChange: number;
  netProfitChange: number;
  jobCountChange: number;
}

export interface DailyRevenue {
  day: string;
  revenue: number;
}

export interface ServiceRevenue {
  serviceType: string;
  revenue: number;
}

export interface TopClient {
  clientName: string;
  totalSpent: number;
  jobCount: number;
}

export interface Job {
  id: number;
  /** @nullable */
  clientId?: number | null;
  date: string;
  clientName: string;
  serviceType: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  location?: string | null;
  amount: number;
  teamMembers: number;
  wages: number;
  netIncome: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface JobInput {
  /** @nullable */
  clientId?: number | null;
  date: string;
  clientName: string;
  serviceType: string;
  description?: string;
  location?: string;
  amount: number;
  teamMembers: number;
  notes?: string;
}

export interface JobUpdate {
  /** @nullable */
  clientId?: number | null;
  date?: string;
  clientName?: string;
  serviceType?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  location?: string | null;
  amount?: number;
  teamMembers?: number;
  /** @nullable */
  notes?: string | null;
}

export interface JobImportRow {
  date: string;
  clientName: string;
  serviceType: string;
  description?: string;
  location?: string;
  amount: number;
  teamMembers: number;
  notes?: string;
}

export interface JobImport {
  rows: JobImportRow[];
}

export interface ImportResult {
  imported: number;
  errors: number;
}

export interface Expense {
  id: number;
  date: string;
  category: string;
  description: string;
  amount: number;
  createdAt: string;
}

export interface ExpenseInput {
  date: string;
  category: string;
  description: string;
  amount: number;
}

export interface ExpenseUpdate {
  date?: string;
  category?: string;
  description?: string;
  amount?: number;
}

export interface ExpenseImportRow {
  date: string;
  category: string;
  description: string;
  amount: number;
}

export interface ExpenseImport {
  rows: ExpenseImportRow[];
}

export interface ExpenseCategorySummary {
  category: string;
  total: number;
}

export interface Client {
  id: number;
  clientCode: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  location?: string | null;
  status: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  firstVisitDate?: string | null;
  createdAt: string;
}

export interface ClientInput {
  name: string;
  phone?: string;
  email?: string;
  location?: string;
  status: string;
  notes?: string;
  firstVisitDate?: string;
}

export interface ClientUpdate {
  name?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  location?: string | null;
  status?: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  firstVisitDate?: string | null;
}

export interface ClientStats {
  totalSpent: number;
  totalVisits: number;
  averageSpend: number;
  /** @nullable */
  favouriteService: string | null;
  /** @nullable */
  firstVisitDate?: string | null;
  /** @nullable */
  lastVisitDate?: string | null;
}

export interface ClientTransaction {
  id: number;
  date: string;
  serviceType: string;
  /** @nullable */
  description?: string | null;
  amount: number;
  netIncome: number;
}

export interface ClientProfile {
  client: Client;
  stats: ClientStats;
  transactions: ClientTransaction[];
}

export interface Receipt {
  id: number;
  receiptNumber: string;
  /** @nullable */
  jobId?: number | null;
  clientName: string;
  serviceType: string;
  /** @nullable */
  description?: string | null;
  amount: number;
  date?: string;
  createdAt: string;
}

export interface ReceiptInput {
  /** @nullable */
  jobId?: number | null;
  clientName: string;
  serviceType: string;
  description?: string;
  amount: number;
  date: string;
}

export interface PLMonth {
  month: string;
  revenue: number;
  expenses: number;
  netProfit: number;
  jobCount: number;
  /** @nullable */
  revenueChange?: number | null;
}

export interface MonthlyTrend {
  month: string;
  revenue: number;
}

export interface KeyStats {
  bestMonthByRevenue: string;
  bestMonthByProfit: string;
  mostPopularService: string;
  topClientAllTime: string;
  avgRevenuePerJob: number;
  avgMonthlyProfit: number;
}

export interface MonthDrill {
  month: string;
  revenue: number;
  expenses: number;
  netProfit: number;
  jobCount: number;
  jobs: Job[];
  expenseBreakdown: ExpenseCategorySummary[];
}

export interface Settings {
  wagePerPersonPerDay: number;
  monthlyRent: number;
}

export interface SettingsUpdate {
  wagePerPersonPerDay?: number;
  monthlyRent?: number;
}

export type GetDashboardStatsParams = {
/**
 * Month in YYYY-MM format (defaults to current month)
 */
month?: string;
};

export type GetDailyRevenueParams = {
/**
 * Month in YYYY-MM format
 */
month?: string;
};

export type GetRevenueByServiceParams = {
month?: string;
};

export type GetTopClientsParams = {
month?: string;
};

export type ListJobsParams = {
/**
 * Filter by month YYYY-MM
 */
month?: string;
serviceType?: string;
};

export type ListExpensesParams = {
month?: string;
category?: string;
};

export type GetExpensesMonthlySummaryParams = {
month: string;
};

export type ListClientsParams = {
search?: string;
status?: string;
};

export type GetMonthDrillParams = {
month: string;
};

export type ExportData200 = { [key: string]: unknown };