# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gold Standard Cleaners internal management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
  /auth/password:
    patch:
      operationId: updatePassword
      tags: [auth]
      summary: Change admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordUpdate"
      responses:
        "200":
          description: Password updated
        "400":
          description: Bad request

  # ─── DASHBOARD ────────────────────────────────────────────────────────────
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard summary stats for a given month
      parameters:
        - in: query
          name: month
          schema:
            type: string
          description: "Month in YYYY-MM format (defaults to current month)"
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/daily-revenue:
    get:
      operationId: getDailyRevenue
      tags: [dashboard]
      summary: Daily revenue for current month (bar chart data)
      parameters:
        - in: query
          name: month
          schema:
            type: string
          description: "Month in YYYY-MM format"
      responses:
        "200":
          description: Daily revenue array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyRevenue"
  /dashboard/revenue-by-service:
    get:
      operationId: getRevenueByService
      tags: [dashboard]
      summary: Revenue by service type (donut chart data)
      parameters:
        - in: query
          name: month
          schema:
            type: string
      responses:
        "200":
          description: Revenue by service
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceRevenue"
  /dashboard/top-clients:
    get:
      operationId: getTopClients
      tags: [dashboard]
      summary: Top 5 clients by amount this month
      parameters:
        - in: query
          name: month
          schema:
            type: string
      responses:
        "200":
          description: Top clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopClient"
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Add a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client profile with stats and transaction history
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientProfile"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update client details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── RECEIPTS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    PasswordUpdate:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string }
    AuthUser:
      type: object
      required: [id, username]
      properties:
        id: { type: integer }
        username: { type: string }

    # ── Dashboard ──
    DashboardStats:
      type: object
      required: [revenue, expenses, netProfit, jobCount, revenueChange, expensesChange, netProfitChange, jobCountChange]
      properties:
        revenue: { type: number }
        expenses: { type: number }
        netProfit: { type: number }
        jobCount: { type: integer }
        revenueChange: { type: number }
        expensesChange: { type: number }
        netProfitChange: { type: number }
        jobCountChange: { type: number }
    DailyRevenue:
      type: object
      required: [day, revenue]
      properties:
        day: { type: string }
        revenue: { type: number }
    ServiceRevenue:
      type: object
      required: [serviceType, revenue]
      properties:
        serviceType: { type: string }
        revenue: { type: number }
    TopClient:
      type: object
      required: [clientName, totalSpent, jobCount]
      properties:
        clientName: { type: string }
        totalSpent: { type: number }
        jobCount: { type: integer }

    # ── Jobs ──
...
    Client:
      type: object
      required: [id, clientCode, name, status, createdAt]
      properties:
        id: { type: integer }
        clientCode: { type: string }
        name: { type: string }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        location: { type: ["string", "null"] }
        status: { type: string }
        notes: { type: ["string", "null"] }
        firstVisitDate: { type: ["string", "null"] }
        createdAt: { type: string }
    ClientInput:
      type: object
      required: [name, status]
      properties:
        name: { type: string }
        phone: { type: string }
        email: { type: string }
        location: { type: string }
        status: { type: string }
        notes: { type: string }
        firstVisitDate: { type: string }
    ClientUpdate:
      type: object
      properties:
        name: { type: string }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        location: { type: ["string", "null"] }
        status: { type: string }
        notes: { type: ["string", "null"] }
        firstVisitDate: { type: ["string", "null"] }
    ClientStats:
      type: object
      required: [totalSpent, totalVisits, averageSpend, favouriteService]
      properties:
        totalSpent: { type: number }
        totalVisits: { type: integer }
        averageSpend: { type: number }
        favouriteService: { type: ["string", "null"] }
        firstVisitDate: { type: ["string", "null"] }
        lastVisitDate: { type: ["string", "null"] }
    ClientTransaction:
      type: object
      required: [id, date, serviceType, amount, netIncome]
      properties:
        id: { type: integer }
        date: { type: string }
        serviceType: { type: string }
        description: { type: ["string", "null"] }
        amount: { type: number }
        netIncome: { type: number }
    ClientProfile:
      type: object
      required: [client, stats, transactions]
      properties:
        client:
          $ref: "#/components/schemas/Client"
        stats:
          $ref: "#/components/schemas/ClientStats"
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/ClientTransaction"

    # ── Receipts ──
...