# lib/api-zod/src/generated/types/clientStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gold Standard Cleaners internal management API
 * OpenAPI spec version: 0.1.0
 */

export interface ClientStats {
  totalSpent: number;
  totalVisits: number;
  averageSpend: number;
  /** @nullable */
  favouriteService: string | null;
  /** @nullable */
  firstVisitDate?: string | null;
  /** @nullable */
  lastVisitDate?: string | null;
}