# lib/api-zod/src/generated/types/topClient.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gold Standard Cleaners internal management API
 * OpenAPI spec version: 0.1.0
 */

export interface TopClient {
  clientName: string;
  totalSpent: number;
  jobCount: number;
}